+++ tests/test_scopes.py
@@ -2,6 +2,7 @@
 """Test the Scopes."""
 
 import os
+import pathlib
 import subprocess
 import sys
 
@@ -14,7 +15,7 @@ import pyvisa as visa
 from packaging.version import Version
 
 from tm_devices import DeviceManager
-from tm_devices.drivers import MSO2, MSO5, MSO5B, MSO6, MSO70KDX, TekScopeSW
+from tm_devices.drivers import MSO2, MSO2KB, MSO5, MSO5B, MSO6, MSO70KDX, TekScopeSW
 from tm_devices.drivers.pi.scopes.tekscope.tekscope import (
     ExtendedSourceDeviceConstants,
     ParameterBounds,
@@ -482,3 +483,44 @@ def test_tekscopesw(device_manager: DeviceManager) -> None:
     assert scope.ip_address == ""
     assert not scope.total_channels
     assert scope.all_channel_names_list == ()
+
+
+def test_tekscope2k(device_manager: DeviceManager, tmp_path: pathlib.Path) -> None:
+    """Test the tekscope2k implementation.
+
+    Args:
+        device_manager: The DeviceManager object.
+        tmp_path: pytest temporary directory function
+    """
+    device_manager.remove_all_devices()
+    # Verify hostname can be determined when only the IP is provided
+    scope: MSO2KB = device_manager.add_scope(
+        "MSO2KB-SERIAL1", alias="mso2kb", connection_type="USB"
+    )
+
+    assert scope.idn_string == "TEKTRONIX,MSO2024B,001061,CF:91.1CT FV:v1.30"
+
+    assert scope.all_channel_names_list == ("CH1", "CH2", "CH3", "CH4")
+    assert scope.total_channels == 4
+
+    # Test toggle all channels on/off
+    scope.turn_all_channels_on()
+    for ch_name in scope.all_channel_names_list:
+        assert scope.query(f"SELECT:{ch_name}?") == "1"
+
+    scope.turn_all_channels_off()
+    for ch_name in scope.all_channel_names_list:
+        assert scope.query(f"SELECT:{ch_name}?") == "0"
+
+    filename = tmp_path / "temp.txt"
+    curve = scope.curve_query(1, wfm_type="TimeDomain", output_csv_file=str(filename))
+    with open(filename, encoding="utf8") as curve_csv:
+        curve_reformatted_from_file = list(map(float, curve_csv.read()[:-1].split(",")))
+        assert curve == curve_reformatted_from_file
+
+    with pytest.raises(AssertionError):
+        scope.curve_query(5, wfm_type="FreqDomain")
+
+    assert scope.curve_query(5, wfm_type="TimeDomain") == []
+
+    assert scope.curve_query(0, wfm_type="Digital") == [1, 0, 1, 0, 1]