+++ CHANGELOG.md
@@ -18,6 +18,10 @@ Valid subsections within a version are:
 
 Things to be included in the next release go here.
 
+### Added
+
+- Added curve query support for the MSO2KB series scopes
+
 ---
 
 ## v2.2.2 (2024-08-14)